Wait for Marathon to restart in RestartIntegrationTest.
Summary:
This explains some falkiness. We do not wait until Marathon is up again.
The test failed 1 in 2 times on my machine. With this fix it does not.

Remove unstable test targets.

We remove the unstable targets and define three kinds of tests:
  * Unit tests that run in parallel on a forked JVM.
  * Integration tests that run in parallel on a forked JVM.
  * Serial integration tests that run in sequence after the integration
    tests.

Integration tests should run in parallel. Only if they are considered
flaky they should run in sequence. The master pipeline and the
marathon-unstable-loop run the same targets.

Test Plan:
integration-test, loop

pipeline

Reviewers: jasongilanfarr, timcharper, jenkins, zen-dog, aquamatthias, meichstedt, unterstein

Subscribers: jenkins, marathon-team

Differential Revision: https://phabricator.mesosphere.com/D642